% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\name{first_name}
\alias{first_name}
\alias{first_name<-}
\title{Get or set a name's first name}
\usage{
first_name(x)

first_name(x) <- value
}
\arguments{
\item{x}{a name, or vector of names}

\item{value}{a replacement value for x's first name.}
}
\description{
as in the lubridate package, individual components of a name
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a first name
example_name <- "Mr Jim Jeffries"
first_name(example_name)

#Set a first name
first_name(example_name) <- "Prof"
}
\seealso{
\code{\link{salutation}}, \code{\link{middle_name}}, \code{\link{last_name}}
and \code{\link{suffix}} for other accessors.
}

