% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/humaniformat.R
\docType{package}
\name{humaniformat}
\alias{humaniformat}
\alias{humaniformat-package}
\title{A Parser for Human Names}
\description{
Human names are complicated and nonstandard things. Humaniformat attempts to provide functions for parsing those names,
making a best-guess attempt to distinguish sub-components such as prefixes, suffixes, middle names and salutations.
}

