% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\name{middle_name}
\alias{middle_name}
\alias{middle_name<-}
\title{Get or set a name's middle name}
\usage{
middle_name(x)

middle_name(x) <- value
}
\arguments{
\item{x}{a name, or vector of names}

\item{value}{a replacement value for x's middle name.}
}
\description{
as in the lubridate package, individual components of a name
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a middle name
example_name <- "Mr Jim Toby Jeffries"
middle_name(example_name)

#Set a middle name
middle_name(example_name) <- "Richard"
}
\seealso{
\code{\link{salutation}}, \code{\link{first_name}}, \code{\link{last_name}}
and \code{\link{suffix}} for other accessors.
}

