% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{salutation}
\alias{salutation}
\alias{salutation<-}
\title{Get or set a name's saltation}
\usage{
salutation(x)

salutation(x) <- value
}
\arguments{
\item{x}{a name, or vector of names}

\item{value}{a replacement value for x's salutation}
}
\description{
as in the lubridate package, individual components of a name
can be both extracted or set using the relevant function call - see the
examples. In the event that you attempt to set a component to NA, no modification
will be made; in the event that you try to get a component that isn't present, an
NA will be returned.
}
\examples{
#Get a salutation
example_name <- "Mr Jim Jeffries"
salutation(example_name)

#Set a salutation
salutation(example_name) <- "Prof"
}
\seealso{
\code{\link{first_name}}, \code{\link{middle_name}}, \code{\link{last_name}} 
and \code{\link{suffix}} for other accessors.
}

