% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\name{suffix}
\alias{suffix}
\alias{suffix<-}
\title{Get or set a name's suffix}
\usage{
suffix(x)

suffix(x) <- value
}
\arguments{
\item{x}{a name, or vector of names}

\item{value}{a replacement value for x's suffix.}
}
\description{
as in the lubridate package, individual components of a name
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a suffix]
example_name <- "Mr Jim Toby Jeffries Esq"
suffix(example_name)

#Set a suffix
suffix(example_name) <- "PhD"
}
\seealso{
\code{\link{salutation}}, \code{\link{first_name}}, \code{\link{middle_name}}
and \code{\link{last_name}} for other accessors.
}

